+++ src/middlewares/versioned_flat_key_value/pending_part/tree/mod.rs
@@ -1,7 +1,7 @@
-mod add_node_support;
-mod change_root_support;
-mod checkout_support;
-mod key_value_store_manager_support;
+mod add_node_commands;
+mod change_root_commands;
+mod checkout_commands;
+mod key_value_store_manager_commands;
 mod node;
 
 pub type SlabIndex = usize;
